+++ spec/acta/testing_dsl_spec.rb
@@ -30,8 +30,6 @@ RSpec.describe Acta::Testing::DSL, :active_record do
 
   let(:command_class) do
     klass = Class.new(Acta::Command) do
-      stream :book, key: :book_id
-
       param :book_id, :string
       param :new_name, :string
       validates :book_id, :new_name, presence: true